Throw exception when adding or subtracting monies in different contexts
As discussed in the issue tracker, operations involving monies in different contexts may be a sign of a potential defect, even if the result could fit in the context of the left operand. Plus, allowing this behaviour by default would mean that A+B and B+A could return different results (two equal results with different contexts, or one result and one exception).

It is usually better to explicitly juggle contexts if required, or force the operation explicitly using ->getAmount(), as suggested by the exception message.