CurrencyConverter::convert() now supports Money, MoneyBag, and RationalMoney
This was before implemented as two different methods for Money and MoneyBag, and did not support RationalMoney.
To this end, the MoneyContainer interface has been re-introduced, only slightly different: it now returns BigNumber, allowing any kind of Money regardless of their context, as well as MoneyBag instances and even RationalMoney instances.