Cache Money::formatTo()'s NumberFormatter for improved performance
Calling formatTo() many times spends quite some time in NumberFormatter constructor, that performs I/O, according to the profiler.
Because we usually call formatTo() many times in a row for the same locale, we can get a ~2x speedup with this fix.

The best way to tackle this in the future would be to deprecate the formatTo() method and introduce a MoneyFormatter object, that would keep a local cache of all locale/scale pairs.